Fix sketch compilation CI configuration for MKR Zero (#35)
The "Compile Examples" GitHub Actions workflow provides a basic "smoke test" on the project by compiling the appropriate
example sketches for a selection of supported boards on every push, pull request, and periodically.

Some of the examples are written specifically for the Arduino Esplora, with the rest intended for any other supported
board. In order to avoid spurious compilation failures, the workflow is configured to compile the correct examples for
each board by assigning them a "type" attribute. This attribute was omitted for the `arduino:samd:mkrzero` board, which
resulted in no examples being compiled for it.

The correct "type" attribute is hereby assigned to that board and all the non-Esplora example sketches will now be
compiled for it.